Add files via upload
My optimized version for arduino. Added function to get back both values in a struct, reducing the processor time to get both values. Removed moving all values in memory to keep the order. This is not necessary because all values are only summarized later and for that the order is not important. Also for speedup espacially with lots of values.